+++ exercises/exercise_sheet_1/1_3_5.py
@@ -2,35 +2,35 @@ import numpy as np
 from scipy.stats import iqr
 
 
-def mean(values: list) -> int:
+def mean(values: list) -> float:
     return np.mean(values)
 
 
-def median(values: list) -> int:
+def median(values: list) -> float:
     return np.median(values)
 
 
-def mode(values: list) -> int:
+def mode(values: list) -> float:
     return max(set(values), key=values.count)
 
 
-def quantile(values: list, q: float) -> int:
+def quantile(values: list, q: float) -> float:
     return np.quantile(values, q, method="averaged_inverted_cdf")
 
 
-def variance(values: list) -> int:
+def variance(values: list) -> float:
     return np.var(values, ddof=1)
 
 
-def standard_deviation(values: list) -> int:
+def standard_deviation(values: list) -> float:
     return np.std(values, ddof=1)
 
 
-def interquartile_range(values: list) -> int:
+def interquartile_range(values: list) -> float:
     return iqr(values, interpolation="averaged_inverted_cdf")
 
 
-def span(values: list) -> int:
+def span(values: list) -> float:
     max_value = max(values)
     min_value = min(values)
     return max_value - min_value
