+++ exercises/exercise_sheet_1/1_6_1.py
@@ -2,19 +2,19 @@ import numpy as np
 from scipy.stats import iqr
 
 
-def mean(values: list) -> int:
+def mean(values: list) -> float:
     return np.mean(values)
 
 
-def median(values: list) -> int:
+def median(values: list) -> float:
     return np.median(values)
 
 
-def quantile(values: list, q: float) -> int:
+def quantile(values: list, q: float) -> float:
     return np.quantile(values, q, method="averaged_inverted_cdf")
 
 
-def standard_deviation(values: list) -> int:
+def standard_deviation(values: list) -> float:
     return np.std(values, ddof=1)
